requirements.txt
asgiref==3.7.2
Django==5.0.1
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
